working fine"

+++ app/core/config.py
@@ -1,11 +1,14 @@
-from pydantic import BaseSettings
+from pydantic_settings import BaseSettings
+
 
 class Settings(BaseSettings):
     DATABASE_URL: str
     API_KEY: str
-    ALERT_THRESHOLD: int
+    # allow fractional thresholds like 0.75
+    ALERT_THRESHOLD: float
 
     class Config:
         env_file = ".env"
+
 
 settings = Settings()
